Extracted we_unix.h from edit.h
