+++ WeXterm.h
@@ -16,6 +16,9 @@ extern "C" {
 \* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
 #include <X11/Xlib.h>
 #include "Xwpe.h"
+#include "edit.h"
+#include "we_xterm.h"
+//#include "we_unix.h"
 
 /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - *\
   Defines
