+++ unixmakr.h
@@ -66,7 +66,6 @@ extern char *ctree[5];
 #define e_getch() (*e_u_getch)()
 #define fk_putchar(c) (*fk_u_putchar)(c)
 #define e_d_switch_out(c) (*e_u_d_switch_out)(c)
-#define e_switch_screen(sw) (*e_u_switch_screen)(sw)
 #define e_deb_out(f) (*e_u_deb_out)(f)
 #define e_cp_X_to_buffer(f) (*e_u_cp_X_to_buffer)(f)
 #define e_copy_X_buffer(f) (*e_u_copy_X_buffer)(f)
@@ -83,8 +82,6 @@ extern char *ctree[5];
        0   Right shift
 \* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
 #define bioskey() (*u_bioskey)()
-#define e_sys_ini() (*e_u_sys_ini)()
-#define e_sys_end() (*e_u_sys_end)()
 #define e_frb_menue(sw, xa, ya, f, md) (*e_frb_u_menue)(sw, xa, ya, f, md)
 #define e_pr_col_kasten(xa, ya, x, y, f, sw) \
 		(*e_pr_u_col_kasten)(xa, ya, x, y, f, sw)
