+++ we_main.c
@@ -314,7 +314,7 @@ int main(int argc, char **argv)
  }
  ECNT_Init(cn);
  e_ini_unix(&argc, argv);
- e_switch_screen(1);
+ (*e_u_switch_screen)(1);
  fb = e_ini_farbe();
  WpeEditor = cn;
  cn->fb = fb;
